[SofaSimulationCore] Move BaseMechanicalVisitor in its own .h/.cpp files
