+++ SofaKernel/modules/SofaSimulationCore/CMakeLists.txt
@@ -10,6 +10,7 @@ set(HEADER_FILES
     ${SRC_ROOT}/AnimateBeginEvent.h
     ${SRC_ROOT}/AnimateEndEvent.h
     ${SRC_ROOT}/AnimateVisitor.h
+    ${SRC_ROOT}/BaseMechanicalVisitor.h
     ${SRC_ROOT}/BehaviorUpdatePositionVisitor.h
     ${SRC_ROOT}/CactusStackStorage.h
     ${SRC_ROOT}/CleanupVisitor.h
@@ -83,6 +84,7 @@ set(SOURCE_FILES
     ${SRC_ROOT}/AnimateBeginEvent.cpp
     ${SRC_ROOT}/AnimateEndEvent.cpp
     ${SRC_ROOT}/AnimateVisitor.cpp
+    ${SRC_ROOT}/BaseMechanicalVisitor.cpp
     ${SRC_ROOT}/BehaviorUpdatePositionVisitor.cpp
     ${SRC_ROOT}/CactusStackStorage.cpp
     ${SRC_ROOT}/CleanupVisitor.cpp
