+++ SofaKernel/modules/SofaSimulationCore/src/sofa/simulation/MechanicalVisitor.cpp
@@ -38,428 +38,6 @@ namespace simulation
 {
 using namespace sofa::core;
 
-Visitor::Result BaseMechanicalVisitor::processNodeTopDown(simulation::Node* node, VisitorContext* ctx)
-{
-    Result res = RESULT_CONTINUE;
-
-    for (unsigned i=0; i<node->solver.size() && res!=RESULT_PRUNE; i++ )
-    {
-        if(testTags(node->solver[i]))
-        {
-            debug_write_state_before(node->solver[i]);
-            ctime_t t = begin(node, node->solver[i], "fwd");
-            res = this->fwdOdeSolver(ctx, node->solver[i]);
-            end(node, node->solver[i], t);
-            debug_write_state_after(node->solver[i]);
-        }
-    }
-
-
-    if (res != RESULT_PRUNE)
-    {
-
-        if (node->mechanicalState != nullptr)
-        {
-            if (node->mechanicalMapping != nullptr)
-            {
-                if (stopAtMechanicalMapping(node, node->mechanicalMapping))
-                {
-                    // stop all mechanical computations
-                    return RESULT_PRUNE;
-                }
-
-                Result res2 = RESULT_CONTINUE;
-                if(testTags(node->mechanicalMapping))
-                {
-                    debug_write_state_before(node->mechanicalMapping);
-                    ctime_t t = begin(node, node->mechanicalMapping, "fwd");
-                    res = this->fwdMechanicalMapping(ctx, node->mechanicalMapping);
-                    end(node, node->mechanicalMapping , t);
-                    debug_write_state_after(node->mechanicalMapping);
-                }
-
-                if(testTags(node->mechanicalState))
-                {
-                    debug_write_state_before(node->mechanicalState);
-                    ctime_t t = begin(node, node->mechanicalState, "fwd");
-                    res2 = this->fwdMappedMechanicalState(ctx, node->mechanicalState);
-                    end(node, node->mechanicalState, t);
-                    debug_write_state_after(node->mechanicalState);
-                }
-
-                if (res2 == RESULT_PRUNE)
-                    res = res2;
-            }
-            else
-            {
-                if(testTags(node->mechanicalState))
-                {
-                    debug_write_state_before(node->mechanicalState);
-                    ctime_t t = begin(node, node->mechanicalState, "fwd");
-                    res = this->fwdMechanicalState(ctx, node->mechanicalState);
-                    end(node, node->mechanicalState, t);
-                    debug_write_state_after(node->mechanicalState);
-                }
-            }
-        }
-        else if (node->mechanicalMapping != nullptr) // rare case of a mechanical mapping which controls a state located elsewhere.
-        {
-            if (stopAtMechanicalMapping(node, node->mechanicalMapping))
-            {
-                // stop all mechanical computations
-                return RESULT_PRUNE;
-            }
-
-            Result res2 = RESULT_CONTINUE;
-            if(testTags(node->mechanicalMapping))
-            {
-                debug_write_state_before(node->mechanicalMapping);
-                ctime_t t = begin(node, node->mechanicalMapping, "fwd");
-                res = this->fwdMechanicalMapping(ctx, node->mechanicalMapping);
-                end(node, node->mechanicalMapping , t);
-                debug_write_state_after(node->mechanicalMapping);
-            }
-
-            if (res2 == RESULT_PRUNE)
-                res = res2;
-        }
-    }
-
-    if (res != RESULT_PRUNE)
-    {
-        if (node->mass != nullptr)
-        {
-            if(testTags(node->mass))
-            {
-                debug_write_state_before(node->mass);
-                ctime_t t = begin(node, node->mass, "fwd");
-                res = this->fwdMass(ctx, node->mass);
-                end(node, node->mass, t);
-                debug_write_state_after(node->mass);
-            }
-        }
-    }
-
-    if (res != RESULT_PRUNE)
-    {
-        res = for_each_r(this, ctx, node->constraintSolver, &MechanicalVisitor::fwdConstraintSolver);
-    }
-
-    if (res != RESULT_PRUNE)
-    {
-        res = for_each_r(this, ctx, node->forceField, &MechanicalVisitor::fwdForceField);
-    }
-
-    if (res != RESULT_PRUNE)
-    {
-        res = for_each_r(this, ctx, node->interactionForceField, &MechanicalVisitor::fwdInteractionForceField);
-    }
-
-
-    if (res != RESULT_PRUNE)
-    {
-        res = for_each_r(this, ctx, node->projectiveConstraintSet, &MechanicalVisitor::fwdProjectiveConstraintSet);
-    }
-
-    if (res != RESULT_PRUNE)
-    {
-        res = for_each_r(this, ctx, node->constraintSet, &MechanicalVisitor::fwdConstraintSet);
-    }
-
-
-    return res;
-}
-
-
-void BaseMechanicalVisitor::processNodeBottomUp(simulation::Node* node, VisitorContext* ctx)
-{
-    for_each(this, ctx, node->projectiveConstraintSet, &MechanicalVisitor::bwdProjectiveConstraintSet);
-    for_each(this, ctx, node->constraintSet, &MechanicalVisitor::bwdConstraintSet);
-    for_each(this, ctx, node->constraintSolver, &MechanicalVisitor::bwdConstraintSolver);
-
-    if (node->mechanicalState != nullptr)
-    {
-        if (node->mechanicalMapping != nullptr)
-        {
-            if (!stopAtMechanicalMapping(node, node->mechanicalMapping))
-            {
-                if(testTags(node->mechanicalState))
-                {
-                    ctime_t t = begin(node, node->mechanicalState, "bwd");
-                    this->bwdMappedMechanicalState(ctx, node->mechanicalState);
-                    end(node, node->mechanicalState, t);
-                    t = begin(node, node->mechanicalMapping, "bwd");
-                    this->bwdMechanicalMapping(ctx, node->mechanicalMapping);
-                    end(node, node->mechanicalMapping, t);
-                }
-            }
-        }
-        else
-        {
-            if(testTags(node->mechanicalState))
-            {
-                ctime_t t = begin(node, node->mechanicalState, "bwd");
-                this->bwdMechanicalState(ctx, node->mechanicalState);
-                end(node, node->mechanicalState, t);
-            }
-        }
-
-    }
-
-    for (unsigned i=0; i<node->solver.size(); i++ )
-    {
-        if(testTags(node->solver[i]))
-        {
-            ctime_t t = begin(node, node->solver[i], "bwd");
-            this->bwdOdeSolver(ctx, node->solver[i]);
-            end(node, node->solver[i], t);
-        }
-    }
-
-    if (node == root)
-    {
-        root = nullptr;
-    }
-}
-
-
-Visitor::Result BaseMechanicalVisitor::processNodeTopDown(simulation::Node* node)
-{
-    if (root == nullptr)
-    {
-        root = node;
-    }
-
-    VisitorContext ctx;
-    ctx.root = root;
-    ctx.node = node;
-    ctx.nodeData = rootData;
-    return processNodeTopDown(node, &ctx);
-}
-
-
-void BaseMechanicalVisitor::processNodeBottomUp(simulation::Node* node)
-{
-    VisitorContext ctx;
-    ctx.root = root;
-    ctx.node = node;
-    ctx.nodeData = rootData;
-    processNodeBottomUp(node, &ctx);
-}
-
-/// Process all the InteractionConstraint
-Visitor::Result BaseMechanicalVisitor::fwdInteractionProjectiveConstraintSet(VisitorContext* ctx, core::behavior::BaseInteractionProjectiveConstraintSet* c)
-{
-    return fwdProjectiveConstraintSet(ctx->node, c);
-}
-
-/// Process all the InteractionConstraint
-Visitor::Result BaseMechanicalVisitor::fwdInteractionConstraint(VisitorContext* ctx, core::behavior::BaseInteractionConstraint* c)
-{
-    return fwdConstraintSet(ctx->node, c);
-}
-
-
-Visitor::Result BaseMechanicalVisitor::processNodeTopDown(simulation::Node* node, LocalStorage* stack)
-{
-    if (root == nullptr)
-    {
-        root = node;
-    }
-
-    VisitorContext ctx;
-    ctx.root = root;
-    ctx.node = node;
-    ctx.nodeData = rootData;
-
-    const bool writeData = writeNodeData();
-    if (writeData)
-    {
-        // create temporary accumulation buffer for parallel reductions (dot products)
-        if (node != root)
-        {
-            SReal* parentData = stack->empty() ? rootData : (SReal*)stack->top();
-            ctx.nodeData = new SReal(0.0);
-            setNodeData(node, ctx.nodeData, parentData);
-            stack->push(ctx.nodeData);
-        }
-    }
-
-    return processNodeTopDown(node, &ctx);
-}
-
-
-void BaseMechanicalVisitor::processNodeBottomUp(simulation::Node* node, LocalStorage* stack)
-{
-    VisitorContext ctx;
-    ctx.root = root;
-    ctx.node = node;
-    ctx.nodeData = rootData;
-    SReal* parentData = rootData;
-
-    const bool writeData = writeNodeData();
-
-    if (writeData)
-    {
-        // use temporary accumulation buffer for parallel reductions (dot products)
-        if (node != root)
-        {
-            ctx.nodeData = (SReal*)stack->pop();
-            parentData = stack->empty() ? rootData : (SReal*)stack->top();
-        }
-    }
-
-    processNodeBottomUp(node, &ctx);
-
-    if (writeData && parentData != ctx.nodeData)
-        addNodeData(node, parentData, ctx.nodeData);
-}
-
-
-#ifdef SOFA_DUMP_VISITOR_INFO
-
-void BaseMechanicalVisitor::printReadVectors(core::behavior::BaseMechanicalState* mm)
-{
-    if (!mm || !readVector.size() || !Visitor::printActivated || !Visitor::outputStateVector) return;
-
-    printNode("Input");
-    for (unsigned int i=0; i<readVector.size(); ++i)
-    {
-        sofa::core::ConstVecId id = readVector[i].getId(mm);
-        if( ! id.isNull() ) printVector(mm, id );
-    }
-    printCloseNode("Input");
-}
-
-
-void BaseMechanicalVisitor::printWriteVectors(core::behavior::BaseMechanicalState* mm)
-{
-    if (!mm || !writeVector.size() || !Visitor::printActivated || !Visitor::outputStateVector) return;
-
-    printNode("Output");
-    for (unsigned int i=0; i<writeVector.size(); ++i)
-    {
-        sofa::core::VecId id = writeVector[i].getId(mm);
-        if( ! id.isNull() ) printVector(mm, id );
-    }
-    printCloseNode("Output");
-}
-
-
-void BaseMechanicalVisitor::printReadVectors(simulation::Node* node, core::objectmodel::BaseObject* obj)
-{
-    using sofa::core::behavior::BaseInteractionForceField;
-    using sofa::core::behavior::BaseInteractionProjectiveConstraintSet;
-    using sofa::core::behavior::BaseInteractionConstraint;
-
-    if (!Visitor::printActivated || !Visitor::outputStateVector) return;
-
-    if (readVector.size())
-    {
-        core::behavior::BaseMechanicalState *dof1, *dof2;
-
-        if (BaseInteractionForceField* interact = dynamic_cast< BaseInteractionForceField* > (obj))
-        {
-            dof1 = interact->getMechModel1();
-            dof2 = interact->getMechModel2();
-        }
-        else if (BaseInteractionProjectiveConstraintSet* interact = dynamic_cast< BaseInteractionProjectiveConstraintSet* > (obj))
-        {
-            dof1 = interact->getMechModel1();
-            dof2 = interact->getMechModel2();
-        }
-        else if (BaseInteractionConstraint* interact = dynamic_cast< BaseInteractionConstraint* > (obj))
-        {
-            dof1 = interact->getMechModel1();
-            dof2 = interact->getMechModel2();
-        }else
-        {
-            printReadVectors(node->mechanicalState);
-            return;
-        }
-
-        TRACE_ARGUMENT arg1;
-        arg1.push_back(std::make_pair("type", dof1->getClassName()));
-        printNode("Components", dof1->getName(), arg1);
-        printReadVectors(dof1);
-        printCloseNode("Components");
-
-        TRACE_ARGUMENT arg2;
-        arg2.push_back(std::make_pair("type", dof2->getClassName()));
-        printNode("Components", dof2->getName(), arg2);
-        printReadVectors(dof2);
-        printCloseNode("Components");
-    }
-}
-
-
-void BaseMechanicalVisitor::printWriteVectors(simulation::Node* node, core::objectmodel::BaseObject* obj)
-{
-    using sofa::core::behavior::BaseInteractionForceField;
-    using sofa::core::behavior::BaseInteractionProjectiveConstraintSet;
-    using sofa::core::behavior::BaseInteractionConstraint;
-    using sofa::core::behavior::BaseMechanicalState;
-
-    if (!Visitor::printActivated) return;
-
-    if (writeVector.size())
-    {
-        BaseMechanicalState *dof1, *dof2;
-
-        if (BaseInteractionForceField* interact = dynamic_cast< BaseInteractionForceField* > (obj))
-        {
-            dof1 = interact->getMechModel1();
-            dof2 = interact->getMechModel2();
-        }
-        else if (BaseInteractionProjectiveConstraintSet* interact = dynamic_cast< BaseInteractionProjectiveConstraintSet* > (obj))
-        {
-            dof1 = interact->getMechModel1();
-            dof2 = interact->getMechModel2();
-        }
-        else if (BaseInteractionConstraint* interact = dynamic_cast< BaseInteractionConstraint* > (obj))
-        {
-            dof1 = interact->getMechModel1();
-            dof2 = interact->getMechModel2();
-        }else
-        {
-            BaseMechanicalState* dof = node->mechanicalState;
-            if (dof == nullptr)
-                node->getContext()->get(dof);
-            printWriteVectors(dof);
-            return;
-        }
-
-        TRACE_ARGUMENT arg1;
-        arg1.push_back(std::make_pair("type", dof1->getClassName()));
-        printNode("Components", dof1->getName(), arg1);
-        printWriteVectors(dof1);
-        printCloseNode("Components");
-
-        TRACE_ARGUMENT arg2;
-        arg2.push_back(std::make_pair("type", dof2->getClassName()));
-        printNode("Components", dof2->getName(), arg2);
-        printWriteVectors(dof2);
-        printCloseNode("Components");
-    }
-}
-
-
-Visitor::ctime_t BaseMechanicalVisitor::begin(simulation::Node* node, core::objectmodel::BaseObject* obj, const std::string &info)
-{
-    ctime_t t=Visitor::begin(node, obj, info);
-    printReadVectors(node, obj);
-    return t;
-}
-
-
-void BaseMechanicalVisitor::end(simulation::Node* node, core::objectmodel::BaseObject* obj, ctime_t t0)
-{
-    printWriteVectors(node, obj);
-    Visitor::end(node, obj, t0);
-}
-#endif
-
-
 Visitor::Result MechanicalGetDimensionVisitor::fwdMechanicalState(VisitorContext* ctx, core::behavior::BaseMechanicalState* mm)
 {
     const std::size_t n = mm->getMatrixSize();
