+++ SofaKernel/modules/SofaSimulationCore/src/sofa/simulation/MechanicalVisitor.h
@@ -24,7 +24,9 @@
 
 #include <sofa/simulation/Visitor.h>
 #include <sofa/core/VecId.h>
+#include <sofa/core/fwd.h>
 #include <sofa/core/MultiVecId.h>
+#include <sofa/simulation/BaseMechanicalVisitor.h>
 #include <sofa/core/MechanicalParams.h>
 #include <sofa/core/behavior/ForceField.h>
 #include <sofa/core/BaseMapping.h>
@@ -45,337 +47,6 @@ namespace sofa
 
 namespace simulation
 {
-
-/** Base class for easily creating new actions for mechanical simulation.
-
-During the first traversal (top-down), method processNodeTopDown(simulation::Node*) is applied to each simulation::Node.
-Each component attached to this node is processed using the appropriate method, prefixed by fwd.
-During the second traversal (bottom-up), method processNodeBottomUp(simulation::Node*) is applied to each simulation::Node.
-Each component attached to this node is processed using the appropriate method, prefixed by bwd.
-The default behavior of the fwd* and bwd* is to do nothing.
-Derived actions typically overload these methods to implement the desired processing.
-
-*/
-class SOFA_SIMULATION_CORE_API BaseMechanicalVisitor : public Visitor
-{
-
-protected:
-    simulation::Node* root; ///< root node from which the visitor was executed
-    SReal* rootData; ///< data for root node
-
-    virtual Result processNodeTopDown(simulation::Node* node, VisitorContext* ctx);
-    virtual void processNodeBottomUp(simulation::Node* node, VisitorContext* ctx);
-
-public:
-
-    BaseMechanicalVisitor(const sofa::core::ExecParams* params)
-        : Visitor(params)
-		, root(nullptr), rootData(nullptr)
-    {
-		// mechanical visitors shouldn't be able to acess a sleeping node, only visual visitor should
-		canAccessSleepingNode = false;
-	}
-
-    /// Return true if this visitor need to read the node-specific data if given
-    virtual bool readNodeData() const
-    { return false; }
-
-    /// Return true if this visitor need to write to the node-specific data if given
-    virtual bool writeNodeData() const
-    { return false; }
-
-    virtual void setNodeData(simulation::Node* /*node*/, SReal* nodeData, const SReal* parentData)
-    {
-        *nodeData = (parentData == nullptr) ? 0.0 : *parentData;
-    }
-
-    virtual void addNodeData(simulation::Node* /*node*/, SReal* parentData, const SReal* nodeData)
-    {
-        if (parentData)
-            *parentData += *nodeData;
-    }
-
-    static inline void ForceMaskActivate( const helper::vector<sofa::core::behavior::BaseMechanicalState*>& v )
-    {
-        std::for_each( v.begin(), v.end(), [](sofa::core::behavior::BaseMechanicalState* m) {
-            m->forceMask.activate(true);
-        });
-    }
-
-    static inline void ForceMaskDeactivate( const helper::vector<sofa::core::behavior::BaseMechanicalState*>& v)
-    {
-        std::for_each( v.begin(), v.end(), [](sofa::core::behavior::BaseMechanicalState* m) {
-            m->forceMask.activate(false);
-        });
-    }
-
-
-    /// Return a class name for this visitor
-    /// Only used for debugging / profiling purposes
-    const char* getClassName() const override { return "MechanicalVisitor"; }
-
-    /**@name Forward processing
-    Methods called during the forward (top-down) traversal of the data structure.
-    Method processNodeTopDown(simulation::Node*) calls the fwd* methods in the order given here. When there is a mapping, it is processed first, then method fwdMappedMechanicalState is applied to the BaseMechanicalState.
-    When there is no mapping, the BaseMechanicalState is processed first using method fwdMechanicalState.
-    Then, the other fwd* methods are applied in the given order.
-    */
-    ///@{
-
-    /// This method calls the fwd* methods during the forward traversal. You typically do not overload it.
-    Result processNodeTopDown(simulation::Node* node) override;
-
-    /// Parallel version of processNodeTopDown.
-    /// This method calls the fwd* methods during the forward traversal. You typically do not overload it.
-    Result processNodeTopDown(simulation::Node* node, LocalStorage* stack) override;
-
-    /// Process the OdeSolver
-    virtual Result fwdOdeSolver(simulation::Node* /*node*/, sofa::core::behavior::OdeSolver* /*solver*/)
-    {
-        return RESULT_CONTINUE;
-    }
-
-    /// Process the OdeSolver
-    virtual Result fwdOdeSolver(VisitorContext* ctx, sofa::core::behavior::OdeSolver* solver)
-    {
-        return fwdOdeSolver(ctx->node, solver);
-    }
-
-    /// Process the ConstraintSolver
-    virtual Result fwdConstraintSolver(simulation::Node* /*node*/, sofa::core::behavior::ConstraintSolver* /*solver*/)
-    {
-        return RESULT_CONTINUE;
-    }
-
-    /// Process the ConstraintSolver
-    virtual Result fwdConstraintSolver(VisitorContext* ctx, sofa::core::behavior::ConstraintSolver* solver)
-    {
-        return fwdConstraintSolver(ctx->node, solver);
-    }
-
-    /// Process the BaseMechanicalMapping
-    virtual Result fwdMechanicalMapping(simulation::Node* /*node*/, sofa::core::BaseMapping* /*map*/)
-    {
-        return RESULT_CONTINUE;
-    }
-
-    /// Process the BaseMechanicalMapping
-    virtual Result fwdMechanicalMapping(VisitorContext* ctx, sofa::core::BaseMapping* map)
-    {
-        return fwdMechanicalMapping(ctx->node, map);
-    }
-
-    /// Process the BaseMechanicalState if it is mapped from the parent level
-    virtual Result fwdMappedMechanicalState(simulation::Node* /*node*/, sofa::core::behavior::BaseMechanicalState* /*mm*/)
-    {
-        return RESULT_CONTINUE;
-    }
-
-    /// Process the BaseMechanicalState if it is mapped from the parent level
-    virtual Result fwdMappedMechanicalState(VisitorContext* ctx, sofa::core::behavior::BaseMechanicalState* mm)
-    {
-        return fwdMappedMechanicalState(ctx->node, mm);
-    }
-
-    /// Process the BaseMechanicalState if it is not mapped from the parent level
-    virtual Result fwdMechanicalState(simulation::Node* /*node*/, sofa::core::behavior::BaseMechanicalState* /*mm*/)
-    {
-        return RESULT_CONTINUE;
-    }
-
-    /// Process the BaseMechanicalState if it is not mapped from the parent level
-    virtual Result fwdMechanicalState(VisitorContext* ctx, sofa::core::behavior::BaseMechanicalState* mm)
-    {
-        return fwdMechanicalState(ctx->node, mm);
-    }
-
-    /// Process the BaseMass
-    virtual Result fwdMass(simulation::Node* /*node*/, sofa::core::behavior::BaseMass* /*mass*/)
-    {
-        return RESULT_CONTINUE;
-    }
-
-    /// Process the BaseMass
-    virtual Result fwdMass(VisitorContext* ctx, sofa::core::behavior::BaseMass* mass)
-    {
-        return fwdMass(ctx->node, mass);
-    }
-
-    /// Process all the BaseForceField
-    virtual Result fwdForceField(simulation::Node* /*node*/, sofa::core::behavior::BaseForceField* /*ff*/)
-    {
-        return RESULT_CONTINUE;
-    }
-
-
-    /// Process all the BaseForceField
-    virtual Result fwdForceField(VisitorContext* ctx, sofa::core::behavior::BaseForceField* ff)
-    {
-        return fwdForceField(ctx->node, ff);
-    }
-
-    /// Process all the InteractionForceField
-    virtual Result fwdInteractionForceField(simulation::Node* node, sofa::core::behavior::BaseInteractionForceField* ff)
-    {
-        return fwdForceField(node, ff);
-    }
-
-    /// Process all the InteractionForceField
-    virtual Result fwdInteractionForceField(VisitorContext* ctx, sofa::core::behavior::BaseInteractionForceField* ff)
-    {
-        return fwdInteractionForceField(ctx->node, ff);
-    }
-
-    /// Process all the BaseProjectiveConstraintSet
-    virtual Result fwdProjectiveConstraintSet(simulation::Node* /*node*/, sofa::core::behavior::BaseProjectiveConstraintSet* /*c*/)
-    {
-        return RESULT_CONTINUE;
-    }
-
-    /// Process all the BaseConstraintSet
-    virtual Result fwdConstraintSet(simulation::Node* /*node*/,sofa::core::behavior::BaseConstraintSet* /*c*/)
-    {
-        return RESULT_CONTINUE;
-    }
-
-    /// Process all the BaseProjectiveConstraintSet
-    virtual Result fwdProjectiveConstraintSet(VisitorContext* ctx,sofa::core::behavior::BaseProjectiveConstraintSet* c)
-    {
-        return fwdProjectiveConstraintSet(ctx->node, c);
-    }
-
-    /// Process all the BaseConstraintSet
-    virtual Result fwdConstraintSet(VisitorContext* ctx,sofa::core::behavior::BaseConstraintSet* c)
-    {
-        return fwdConstraintSet(ctx->node, c);
-    }
-
-    /// Process all the InteractionConstraint
-    virtual Result fwdInteractionProjectiveConstraintSet(simulation::Node* /*node*/,sofa::core::behavior::BaseInteractionProjectiveConstraintSet* /*c*/)
-    {
-        return RESULT_CONTINUE;
-    }
-
-    /// Process all the InteractionConstraint
-    virtual Result fwdInteractionConstraint(simulation::Node* /*node*/,sofa::core::behavior::BaseInteractionConstraint* /*c*/)
-    {
-        return RESULT_CONTINUE;
-    }
-
-    /// Process all the InteractionConstraint
-    virtual Result fwdInteractionProjectiveConstraintSet(VisitorContext* ctx,sofa::core::behavior::BaseInteractionProjectiveConstraintSet* c);
-
-    /// Process all the InteractionConstraint
-    virtual Result fwdInteractionConstraint(VisitorContext* ctx,sofa::core::behavior::BaseInteractionConstraint* c);
-
-    ///@}
-
-    /**@name Backward processing
-    Methods called during the backward (bottom-up) traversal of the data structure.
-    Method processNodeBottomUp(simulation::Node*) calls the bwd* methods.
-    When there is a mapping, method bwdMappedMechanicalState is applied to the BaseMechanicalState.
-    When there is no mapping, the BaseMechanicalState is processed using method bwdMechanicalState.
-    Finally, the mapping (if any) is processed using method bwdMechanicalMapping.
-    */
-    ///@{
-
-    /// This method calls the bwd* methods during the backward traversal. You typically do not overload it.
-    void processNodeBottomUp(simulation::Node* node) override;
-
-    /// Parallel version of processNodeBottomUp.
-    /// This method calls the bwd* methods during the backward traversal. You typically do not overload it.
-    void processNodeBottomUp(simulation::Node* /*node*/, LocalStorage* stack) override;
-
-    /// Process the BaseMechanicalState when it is not mapped from parent level
-    virtual void bwdMechanicalState(simulation::Node* /*node*/,sofa::core::behavior::BaseMechanicalState* /*mm*/)
-    {}
-
-    /// Process the BaseMechanicalState when it is not mapped from parent level
-    virtual void bwdMechanicalState(VisitorContext* ctx,sofa::core::behavior::BaseMechanicalState* mm)
-    { bwdMechanicalState(ctx->node, mm); }
-
-    /// Process the BaseMechanicalState when it is mapped from parent level
-    virtual void bwdMappedMechanicalState(simulation::Node* /*node*/,sofa::core::behavior::BaseMechanicalState* /*mm*/)
-    {}
-
-    /// Process the BaseMechanicalState when it is mapped from parent level
-    virtual void bwdMappedMechanicalState(VisitorContext* ctx,sofa::core::behavior::BaseMechanicalState* mm)
-    { bwdMappedMechanicalState(ctx->node, mm); }
-
-    /// Process the BaseMechanicalMapping
-    virtual void bwdMechanicalMapping(simulation::Node* /*node*/, sofa::core::BaseMapping* /*map*/)
-    {}
-
-    /// Process the BaseMechanicalMapping
-    virtual void bwdMechanicalMapping(VisitorContext* ctx, sofa::core::BaseMapping* map)
-    { bwdMechanicalMapping(ctx->node, map); }
-
-    /// Process the OdeSolver
-    virtual void bwdOdeSolver(simulation::Node* /*node*/,sofa::core::behavior::OdeSolver* /*solver*/)
-    {}
-
-    /// Process the OdeSolver
-    virtual void bwdOdeSolver(VisitorContext* ctx,sofa::core::behavior::OdeSolver* solver)
-    { bwdOdeSolver(ctx->node, solver); }
-
-    /// Process the ConstraintSolver
-    virtual void bwdConstraintSolver(simulation::Node* /*node*/,sofa::core::behavior::ConstraintSolver* /*solver*/)
-    {}
-
-    /// Process the ConstraintSolver
-    virtual void bwdConstraintSolver(VisitorContext* ctx,sofa::core::behavior::ConstraintSolver* solver)
-    { bwdConstraintSolver(ctx->node, solver); }
-
-    /// Process all the BaseProjectiveConstraintSet
-    virtual void bwdProjectiveConstraintSet(simulation::Node* /*node*/,sofa::core::behavior::BaseProjectiveConstraintSet* /*c*/)
-    {}
-
-    /// Process all the BaseConstraintSet
-    virtual void bwdConstraintSet(simulation::Node* /*node*/,sofa::core::behavior::BaseConstraintSet* /*c*/)
-    {}
-
-    /// Process all the BaseProjectiveConstraintSet
-    virtual void bwdProjectiveConstraintSet(VisitorContext* ctx,sofa::core::behavior::BaseProjectiveConstraintSet* c)
-    { bwdProjectiveConstraintSet(ctx->node, c); }
-
-    /// Process all the BaseConstraintSet
-    virtual void bwdConstraintSet(VisitorContext* ctx,sofa::core::behavior::BaseConstraintSet* c)
-    { bwdConstraintSet(ctx->node, c); }
-
-    ///@}
-
-
-    /// Return a category name for this action.
-    /// Only used for debugging / profiling purposes
-    const char* getCategoryName() const override
-    {
-        return "animate";
-    }
-
-    virtual bool stopAtMechanicalMapping(simulation::Node* /*node*/, sofa::core::BaseMapping* map)
-    {
-        return !map->areForcesMapped();
-    }
-
-#ifdef SOFA_DUMP_VISITOR_INFO
-    ctime_t begin(simulation::Node* node, sofa::core::objectmodel::BaseObject* obj, const std::string &info=std::string("type")) override;
-    void end(simulation::Node* node, sofa::core::objectmodel::BaseObject* obj, ctime_t t0) override;
-#endif
-
-#ifdef SOFA_DUMP_VISITOR_INFO
-    virtual void setReadWriteVectors() {}
-    virtual void addReadVector(core::ConstMultiVecId id) {  readVector.push_back(id);  }
-    virtual void addWriteVector(core::MultiVecId id) {  writeVector.push_back(id);  }
-    virtual void addReadWriteVector(core::MultiVecId id) {  readVector.push_back(core::ConstMultiVecId(id)); writeVector.push_back(id);  }
-    void printReadVectors(core::behavior::BaseMechanicalState* mm);
-    void printReadVectors(simulation::Node* node, sofa::core::objectmodel::BaseObject* obj);
-    void printWriteVectors(core::behavior::BaseMechanicalState* mm);
-    void printWriteVectors(simulation::Node* node, sofa::core::objectmodel::BaseObject* obj);
-protected:
-    sofa::helper::vector< sofa::core::ConstMultiVecId > readVector;
-    sofa::helper::vector< sofa::core::MultiVecId > writeVector;
-#endif
-};
 
 class SOFA_SIMULATION_CORE_API MechanicalVisitor : public BaseMechanicalVisitor
 {
